test-api-main/requirements.txt
asgiref==3.3.1
attrs==21.4.0
boto3==1.12.34
botocore==1.15.49
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.9
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.8
django-filter==2.4.0
django-object-actions==3.0.2
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==4.6.0
docutils==0.15.2
drf-spectacular==0.21.1
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==4.3.2
MarkupSafe==2.0.1
mysqlclient==2.1.0
NGPIris==3.2.0
openapi-codec==1.3.2
packaging==21.3
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
requests==2.25.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.3.7
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.1
uritemplate==4.1.1
urllib3
